fix: disable twitter media preview
https://mask.atlassian.net/browse/FW-1663
